SYSTEM.CSRRx operations
Implement three new SYSTEM standard functions:

  x := SYSTEM.CSRRW(csr, y);
  x := SYSTEM.CSRRS(csr, y);
  x := SYSTEM.CSRRC(csr, y);

csr MUST be an integer constant, or a CONST alias.

y can be any expression that evaluates to an integer or a set.

Set notation is particularly useful for managing interrupts.
Integers are useful for working with addresses.
Either integers or sets can be used for accessing specific sub-fields.

The result of these functions is the *original* value of the CSR.

See RISC-V Privileged Instruction manual for semantics on the CSRRx
instructions.