First step of relocating entry code.
Currently, it still stays in memory like the rest of module code; but
the entry part itself is moved to the current tail of module space, to
be overwritten by the next loaded module. A better solution will require
not loading the module into memory for fixups in the first place, which
is next up.